backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "gunicorn>=23.0.0",
    "janome>=0.5.0",
    "numpy>=2.3.2",
    "pymupdf>=1.26.4",
    "requests>=2.32.5",
    "ruff>=0.12.12",
    "sentence-transformers>=5.1.0",
    "uvicorn[standard]>=0.35.0",
]
